requirements.txt
requests
openai
python-dotenv
fastapi
uvicorn
jinja2
